Eliminate type header
Previously automatically serialized types (Write (object)) would include
their type names in a "type header" once with an ID#. This was to optimize
repeat uses of a type in the same messages drastically shrinking the
message size. Tempest's internal protocol itself used this serializing RSA
public keys.

Since serializing just any object is no longer supported, this is really
no longer neccessary, it just complicates the message
serialization/deserialization process.

As a side effect, the exchanged public keys in the handshake must no
longer rely on this mechanism and instead now have a defined
serialization/deserialization process. This has the side benefit of no
longer requiring a .NET type to actually connect with a Tempest end point.

Also included in this change set is removing any code that might
potentially serialize a private key. If this behavior is needed, it will
have to be implemented manually.