Fix rapid connect attempt race condition
If you rapidly tried to reconnect, some situations might null out the
connection task you added because the new connect didn't properly block
against a fully completed disconnection. Two changes go into fixing this.

Now, on a failed connect we'll swap out the task before we decrement. This
just moves the swap to before to when the Connect is released from pending
to go ahead with its attempt.

The second change is largely a consolidation of some code. Originally we'd
have a pending async for a real disconnect attempt and not for a "we're
already not connected, just cleanup" disconnect. The later would still
erase the connect task though, so now it all happens the same way to avoid
this situation.

I do have some concerns about the Disconnect() after a failed connect but
not behind a pending async, but that's something I'll have to look at
later.