Make connect/disconnect more reliable
This could be taken a bit further for even more reliability making the
queue and running flag completely isolated per setup of a runner.
Probably moving the flag to a cancel token.

For now this should solve the bug in question. The client was not
expecting the underlying conneciton to reconnect itself so some state
setup was in the client's connection API, rather than the internal
on-connected handler.